Confirmation de la suppression de parking

+++ parkings.php
@@ -32,7 +32,7 @@ if(isset($_GET['id'])){
         <button type="submit" class="btn btn-primary">Enregistrer</button>
         <a class="btn" href="parkings.php">Annuler</a>
         <?php if($id > 0): ?>
-        <a class="btn btn-danger" href="parkings_control.php?delete=<?php echo $id; ?>">Supprimer</a>
+        <a class="btn btn-danger" onclick="return confirm('Voulez-vous vraiment supprimer ce parking ?');" href="parkings_control.php?delete=<?php echo $id; ?>">Supprimer</a>
         <?php endif; ?>
       </div>
     </fieldset>
